[procmgrd] Extract ProcessManager into dedicated manager module
Move ProcessManager, ExitEvent, spawn_watcher, and the config/startup
helper functions out of main.rs into manager.rs. The main.rs now only
contains the tokio main loop and signal handling, making it easier to
navigate the codebase.
