+++ pkg/procmgr/rust/src/grpc/service.rs
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2026-present Datadog, Inc.
 
-use crate::ProcessManager;
+use crate::manager::ProcessManager;
 use crate::command::Command;
 use crate::config::{ProcessConfig, RestartPolicy};
 use crate::grpc::proto;
